Move CanSelectContent to upper component
