+++ src/components/sidebar/CallTreeSidebar.js
@@ -5,7 +5,6 @@
 // @flow
 
 import * as React from 'react';
-import classNames from 'classnames';
 
 import explicitConnect from '../../utils/connect';
 import {
@@ -15,6 +14,7 @@ import {
 import { getSelectedThreadIndex } from '../../reducers/url-state';
 import { getFunctionName } from '../../profile-logic/function-info';
 import { assertExhaustiveCheck } from '../../utils/flow';
+import { CanSelectContent } from './index';
 
 import type {
   ConnectedProps,
@@ -31,44 +31,6 @@ import type {
   StackImplementation,
   TimingsForPath,
 } from '../../profile-logic/profile-data';
-
-type CanCopyContentProps = {|
-  +tagName?: string,
-  +content: string,
-  +className?: string,
-|};
-
-class CanSelectContent extends React.PureComponent<CanCopyContentProps> {
-  _selectContent(e: SyntheticMouseEvent<HTMLInputElement>) {
-    const input = e.currentTarget;
-    input.focus();
-    input.select();
-  }
-
-  _unselectContent(e: SyntheticMouseEvent<HTMLInputElement>) {
-    e.currentTarget.setSelectionRange(0, 0);
-  }
-
-  render() {
-    const { tagName, content, className } = this.props;
-    const TagName = tagName || 'div';
-
-    return (
-      <TagName
-        className={classNames(className, 'can-select-content')}
-        title={`${content}\n(click to select)`}
-      >
-        <input
-          value={content}
-          className="can-select-content-input"
-          onFocus={this._selectContent}
-          onBlur={this._unselectContent}
-          readOnly={true}
-        />
-      </TagName>
-    );
-  }
-}
 
 type SidebarDetailProps = {|
   +label: string,
